py/pyproject.toml
[tool.poetry]
name = "poem-download"
version = "0.1.0"
description = ""
authors = ["NitrogenReaction <nitrogenreaction@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
gspread = "^6.0.0"
google-api-python-client = "^2.116.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
